added a utils mod to the cli
