+++ cli/src/lib.rs
@@ -3,3 +3,5 @@ pub type Error = Box<dyn std::error::Error + Send + Sync + 'static>;
 
 mod actions;
 pub use actions::*;
+mod utils;
+pub use utils::*;
